After realizing how simple it would be to implement last night, I've converted the decay prediction algorithm to a flow based model.  This simply takes difference between the incoming nuclei from parent decays and the difference between the start and end quantities of the nuclei in question and what is left is the amount of decays.
This algorithm is simple to understand and effective.  Passing
verifications 1 and 2 with nearly passing colors.  In verification 1
there seems to be a problem with the higher resolution. A sort of
jittering seems to develop and it causes the value to stray high.  Will
look into that further and see if the onset of this jitter is
predictable, such that a warning can be provided when the wrong mix of
time difference, half-lives, and resolutions are provided.