Added command to retrieve coincidental particle outputs at arbitrary times, as well as, a command "read batch <file>" to read commands in from batch files.
Updated the entire U238 decay chain with more complete data including
coincidentals and recalculated the half-lives based on the nuDat's use
of 1 year = 365.242 days.

Using the new U238 input file, the program does not seem to be properly
compiling and outputting all of the betas, gammas, and alphas that
should exit.  Will have to get to the bottom of that.